requirements.txt
tornado
pycurl
pyopenssl
kafka-python
sqlalchemy